# httpgrpc: reuse response body buffers