@bboreham 2022-10-14
Thanks, this is a clever idea.

> To safely return body buffers back to the pool, we resorted to gRPC stats.Handler to determine whether responses have been effectively serialized and sent over the wire.

Do you have a reference for what makes this safe?  E.g. is it stated in the gRPC documentation?