@ortuman 2022-10-14
> Do you have a reference for what makes this safe? E.g. is it stated in the gRPC documentation?

There's no explicit statement about it (at least I haven't found it). However, the [stats.OutPayload](https://github.com/grpc/grpc-go/blob/v1.50.0/stats/stats.go#L125) has two additional fields that indicate the length of the content that was sent over the wire as well as the timestamp at which the sent took place (which BTW, leads me to think that we could validate if `SentTime` field is not zero before recycling the body buffer, just in case).