@bboreham 2022-10-14
By my reading of the code, this only applies to server responses.  Are there many cases where responses sent by httpgrpc are a sizeable portion of garbage?

Is it possible to apply the same technique to requests, on client or server side?  I believe those are much more of a problem.